refactor(memberInfo): 네이밍 명확히
