+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { ThemeProvider } from "@emotion/react";
 import { QueryClient, QueryClientProvider } from "react-query";
-import { MemberInfoContext } from "~/contexts";
+import { MemberInfo } from "~/contexts";
 import { emotionTheme } from "~/theme";
 
 const queryClient = new QueryClient();
@@ -11,9 +11,9 @@ function MyApp({ Component, pageProps }: AppProps) {
   return (
     <QueryClientProvider client={queryClient}>
       <ThemeProvider theme={emotionTheme}>
-        <MemberInfoContext.Provider>
+        <MemberInfo.Provider>
           <Component {...pageProps} />
-        </MemberInfoContext.Provider>
+        </MemberInfo.Provider>
       </ThemeProvider>
     </QueryClientProvider>
   );
